+++ public/css/users_login.css
@@ -1,22 +1,3 @@
- 
-
-*,
-:before,
-:after {
-  box-sizing: border-box;
-}
-
-html,
-body {
-  height: 100%;
-  width: 100%;
-}
-
-body {
-  font-family: "Roboto Slab", serif;
-  font-size: 10px;
-}
-
 .form-body {
   background-color: #fc9d9a;
   background: linear-gradient(
@@ -27,6 +8,9 @@ body {
     ),
     url("../images/users/login_bg.png") center center / cover;
   background-attachment: fixed;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 
 .mb-2r {
@@ -164,9 +148,11 @@ datalist option {
   left: 0%;
 }
 
-.btn {
-  background: #fff;
-  color: #000;
+.btn.btn-submit {
+  background: #6b0073;
+  background: -webkit-linear-gradient(to right bottom, #ec6ead, #6b0073);
+  background: linear-gradient(to right bottom, #ec6ead, #6b0073);
+
   border: none;
   padding: 0.8rem 3rem;
   border-radius: 2rem;
@@ -177,36 +163,14 @@ datalist option {
   box-shadow: 0 1px 3px rgba(236, 110, 173, 0.562),
     0 1px 2px rgba(236, 110, 173, 0.575);
   transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
-}
-.btn:hover {
-  color: #8b8b8b;
-  box-shadow: 0 7px 14px rgba(0, 0, 0, 0.18), 0 5px 5px rgba(0, 0, 0, 0.12);
-}
-.btn.btn-submit {
-  background: #6b0073;
-  background: -webkit-linear-gradient(to right bottom, #ec6ead, #6b0073);
-  background: linear-gradient(
-    to right bottom,
-    #ec6ead,
-    #6b0073
-  ); 
-
   color: #fff;
 }
 .btn.btn-submit:hover {
   background: #6b0073;
-  background: -webkit-linear-gradient(
-    to left top,
-    #ec6ead,
-    #6b0073
-  );  
-  background: linear-gradient(
-    to left top,
-    #ec6ead,
-    #6b0073
-  ); 
-
+  background: -webkit-linear-gradient(to left top, #ec6ead, #6b0073);
+  background: linear-gradient(to left top, #ec6ead, #6b0073);
   color: #deeffd;
+  box-shadow: 0 7px 14px rgba(0, 0, 0, 0.18), 0 5px 5px rgba(0, 0, 0, 0.12);
 }
 .btn-box {
   text-align: center;
