+++ public/users_login.html
@@ -23,11 +23,12 @@
           <div class="row no-gutters form-row">
             <!--left side -->
             <div class="col-md-6 left-img p-0">
-              <h2 class="slogan-text">Knowledge is Power </h2>
+              <h2 class="slogan-text">Knowledge is Power</h2>
 
               <!--  slogan -->
               <p class="slogan-details">
-                “ Dont give up. Obstacles can be overcome through strategy and learning.”
+                “ Dont give up. Obstacles can be overcome through strategy and
+                learning.”
               </p>
 
               <div class="btn-box">
@@ -39,7 +40,7 @@
             <div class="col-md-6 p-0">
               <div class="form-wrapper">
                 <h2 class=" text-center ">
-                 <span class="pinks">Users Login</span> 
+                  Users <span class="pinks">Login</span>
                 </h2>
                 <form>
                   <div class="row">
