+++ views/partials/requiredInput.hbs
@@ -1,5 +1,5 @@
 <div class="form-group position-relative my-4">
-    <input type="{{type}}" class="input inline-input" name="{{name}}" required />
+    <input type="{{type}}" class="{{className}}" name="{{name}}" required />
 
     <span class="highlight"></span>
     <span class="bar"></span>
