+++ views/templates/donate.hbs
@@ -7,11 +7,11 @@
             <form>
                 <div class="row">
                     <div class="col-md-6">
-                        {{> requiredInput name="firstname" type="text" label="First Name"}}
+                        {{> requiredInput name="firstname" type="text" label="First Name" className="input inline-input"}}
                     </div>
                     <div class="col-md-6">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="lastname" type="text" label="Last Name"}}
+                            {{> requiredInput name="lastname" type="text" label="Last Name" className="input inline-input"}}
                         </div>
                     </div>
                 </div>
@@ -19,7 +19,7 @@
                 <div class="row">
                     <div class="col-md-12">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="email" type="email" label="Email"}}
+                            {{> requiredInput name="email" type="email" label="Email" className="input inline-input"}}
                         </div>
                     </div>
                 </div>
@@ -27,17 +27,17 @@
                 <div class="row">
                     <div class="col-md-3">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="country" type="text" label="Country"}}
+                            {{> requiredInput name="country" type="text" label="Country" className="input inline-input"}}
                         </div>
                     </div>
                     <div class="col-md-3">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="state" type="text" label="State"}}
+                            {{> requiredInput name="state" type="text" label="State" className="input inline-input"}}
                         </div>
                     </div>
                     <div class="col-md-3">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="zip" type="text" label="Zip"}}
+                            {{> requiredInput name="zip" type="text" label="Zip" className="input inline-input"}}
                         </div>
                     </div>
                 </div>
@@ -63,12 +63,12 @@
                 <div class="row">
                     <div class="col-md-6">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="nameoncard" type="text" label="Name on card"}}
+                            {{> requiredInput name="nameoncard" type="text" label="Name on card" className="input inline-input"}}
                         </div>
                     </div>
                     <div class="col-md-6">
                         <div class="form-group position-relative my-4">
-                            {{> requiredInput name="cardnum" type="text" label="Credit card number"}}
+                            {{> requiredInput name="cardnum" type="text" label="Credit card number" className="input inline-input"}}
                         </div>
                     </div>
                 </div>
