+++ views/templates/home.hbs
@@ -41,7 +41,7 @@
                             <p class="last-item__text">
                                 Why the Gap ?
                             </p>
-                            <a href="/auth/vol/registration" class="btn join-us">Join with us</a>
+                            <a href="/vol/registration" class="btn join-us">Join with us</a>
                         </div>
                     </div>
                 </div>
@@ -115,7 +115,7 @@
                     Protection Programme.
                 </p>
 
-                <a href="/auth/vol/registration" class="btn donate-btn connect-btn">Connect with us</a>
+                <a href="/vol/registration" class="btn donate-btn connect-btn">Connect with us</a>
             </div>
         </div>
     </div>
@@ -166,7 +166,7 @@
                         organization and the volunteer
                     </p>
                 </div>
-                <a href="/auth/vol/registration" class="btn donate-btn">Become volunteer</a>
+                <a href="/vol/registration" class="btn donate-btn">Become volunteer</a>
             </div>
         </div>
     </div>
