+++ views/templates/userlogin.hbs
@@ -1,0 +1,44 @@
+<div class="form-body w-100 h-100">
+    <div class="container p-0">
+        <div class="row no-gutters">
+            <div class="col-10 m-auto">
+                <div class="row no-gutters form-row">
+                    <!--left side -->
+                    <div class="col-md-6 left-img p-0">
+                        <h2 class="slogan-text">Knowledge is Power</h2>
+
+                        <!--  slogan -->
+                        <p class="slogan-details">
+                            “ Dont give up. Obstacles can be overcome through strategy and
+                            learning.”
+                        </p>
+
+                        <div class="btn-box">
+                            <a class="btn btn-submit" href="/auth/user/registration">Sign Up</a>
+                        </div>
+                    </div>
+
+                    <!--right side  -->
+                    <div class="col-md-6 p-0">
+                        <div class="form-wrapper">
+                            <h2 class=" text-center ">
+                                Users <span class="pinks">Login</span>
+                            </h2>
+                            <form>
+                                <div class="row">
+                                    {{> requiredInput type="text" name="username" label="User name" className="input"}}
+                                    {{> requiredInput type="password" name="password" label="Password" className="input"}}
+                                    <div class="btn-box">
+                                        <button class="btn btn-submit" type="submit">
+                                            Log in
+                                        </button>
+                                    </div>
+                                </div>
+                            </form>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
